services/farmer-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1
redis==5.0.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
cryptography==42.0.5
ruff==0.4.4
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
